Improve autoscaling handling
Motivation
----------
An AZRebalance operation may cause an unexpected instance termination.

Mananging swap-rebalance operations is tricky when you can't terminate
instances with delayed replacement.

Modifications
-------------
Disable the AZRebalance process on the autoscaling group.

Set min and max size on the autoscaling group to a wide range, and let
desired control the number of instances.